image: react to package move
