+++ pkg/image/apiserver/admission/limitrange/admission.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/openshift/origin/pkg/image/internalimageutil"
+	"github.com/openshift/origin/pkg/image/apiserver/internalimageutil"
 
 	"k8s.io/klog"
