+++ pkg/image/apiserver/importer/credentials_test.go
@@ -17,7 +17,7 @@ import (
 )
 
 func TestCredentialsForSecrets(t *testing.T) {
-	data, err := ioutil.ReadFile("../../../test/testdata/image-secrets.json")
+	data, err := ioutil.ReadFile("../../../../test/testdata/image-secrets.json")
 	if err != nil {
 		t.Fatal(err)
 	}
