+++ test/integration/dockerregistryclient_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	dockerregistry "github.com/openshift/origin/pkg/image/importer/dockerv1client"
+	dockerregistry "github.com/openshift/origin/pkg/image/apiserver/importer/dockerv1client"
 )
 
 const (
